Refactored media containers
