+++ src/components/file-previews/ImagePreview/ImagePreview.module.scss
@@ -1,23 +1,6 @@
 .container {
-  position: absolute;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
+  position: relative;
   width: 100%;
   height: 100%;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  overflow: hidden;
-  object-fit: fill;
-}
-
-.image {
-  width: 100%;
-  max-height: 100%;
-  max-width: 100%;
-  object-fit: contain;
 }
