+++ src/components/file-previews/VideoPreview/VideoPreview.module.scss
@@ -2,21 +2,4 @@
   position: relative;
   width: 100%;
   height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.video-wrapper {
-  position: relative;
-  width: 100%;
-  height: 100%;
-}
-
-.player {
-  video {
-    object-fit: contain;
-    width: 100%;
-    height: 100%;
-  }
 }
