+++ src/components/file-previews/components/AudioPlayer/Controls/Controls.module.scss
