+++ src/components/file-previews/components/AudioPlayer/ProgressBar/ProgressBar.jsx
