+++ src/components/file-previews/components/AudioPlayer/ProgressBar/ProgressBar.module.scss
