+++ src/components/file-previews/components/ExcelSnapshotReader/ExcelSnapshotReader.module.scss
@@ -1,0 +1,40 @@
+.container {
+  width: 100%;
+  height: 100%;
+  position: relative;
+}
+
+.excel-viewer {
+  max-width: 100%;
+  max-height: 100%;
+  margin: 0 auto;
+  color: black;
+  overflow: hidden;
+
+  table {
+    border: 1px solid;
+    border-collapse: collapse;
+  }
+
+  table tr:nth-child(even) {
+    background-color: #f2f2f2;
+  }
+
+  table tr:nth-child(odd) {
+    background-color: #ffffff;
+  }
+
+  table td {
+    padding: 5px 10px;
+  }
+}
+
+.no-action-overlay {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  pointer-events: none;
+  user-select: none;
+}
