+++ src/components/file-previews/components/ImageReader/ImageReader.module.scss
@@ -11,5 +11,12 @@
   right: 0;
   bottom: 0;
   overflow: hidden;
+  object-fit: fill;
 }
 
+.image {
+  width: 100%;
+  max-height: 100%;
+  max-width: 100%;
+  object-fit: contain;
+}
