+++ src/components/file-previews/components/PdfSnapshotReader/PdfSnapshotReader.module.scss
@@ -1,0 +1,29 @@
+.page {
+  position: fixed;
+  bottom: 15px;
+  left: 50%;
+  transform: translateX(-50%);
+  z-index: 9999;
+  color: #fff;
+  font-family: 'SF Pro Text';
+  font-size: 14px;
+  font-weight: 500;
+  line-height: 150%;
+  letter-spacing: 0.09px;
+}
+
+.pdf-container {
+  position: relative;
+  pointer-events: none;
+  user-select: none;
+}
+
+.no-action-overlay {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  pointer-events: none;
+  user-select: none;
+}
