+++ src/components/file-previews/components/TxtSnapshotReader/TxtSnapshotReader.module.scss
@@ -1,0 +1,15 @@
+.container {
+  width: 100%;
+  height: 100%;
+  margin: 0 auto;
+  color: black;
+  background-color: #fff;
+  overflow: hidden;
+  padding: 15px;
+
+  pre {
+    width: 100%;
+    white-space: pre-wrap;
+    word-wrap: break-word;
+  }
+}
