+++ src/components/file-previews/components/VideoPlayer/Controls/Controls.module.scss
