+++ src/components/file-previews/components/VideoPlayer/ProgressBar/ProgressBar.jsx
