+++ src/components/file-previews/components/VideoPlayer/ProgressBar/ProgressBar.module.scss
