+++ src/components/file-previews/components/VideoPlayer/VideoPlayer.module.scss
@@ -1,0 +1,22 @@
+.container {
+  position: relative;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.video-wrapper {
+  position: relative;
+  width: 100%;
+  height: 100%;
+}
+
+.player {
+  video {
+    object-fit: contain;
+    width: 100%;
+    height: 100%;
+  }
+}
